+++ API/Collection/TagCollection.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Kaliop\eZMigrationBundle\API\Collection;
+
+/**
+ * @todo add phpdoc to suggest typehinting
+ */
+class TagCollection extends AbstractCollection
+{
+    protected $allowedClass = 'Netgen\TagsBundle\API\Repository\Values\Tags\Tag';
+}