+++ WHATSNEW.md
@@ -3,6 +3,8 @@ Version 3.0.0-beta4
 
 * New: It is now possible to replace the executor and step definition in a BeforeStepExecutionEvent
 
+* New: it is now possible to set values to content fields of type eztags
+
 
 Version 3.0.0-beta3
 ===================